refactor: move modules to a separate directory
