+++ src/app.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UserModule } from './user/user.module';
+import { UserModule } from './modules/user/user.module';
 import { ConfigModule } from '@nestjs/config';
-import { ArtistModule } from './artist/artist.module';
+import { ArtistModule } from './modules/artist/artist.module';
 
 @Module({
   imports: [ConfigModule.forRoot({ isGlobal: true }), UserModule, ArtistModule],
