+++ src/modules/album/album.module.ts
@@ -4,7 +4,7 @@ import { AlbumService } from './album.service';
 import { AlbumRepository } from './album.repository';
 import { TrackModule } from '../track/track.module';
 import { FavoritesModule } from '../favorites/favorites.module';
-import { DatabaseModule } from '../database/database.module';
+import { DatabaseModule } from '../../database/database.module';
 
 @Module({
   imports: [
