+++ src/modules/album/entities/album.entity.ts
