+++ src/modules/artist/entities/artist.entity.ts
