+++ src/modules/favorites/dtos/favorites-response.dto.ts
