+++ src/modules/favorites/entities/favorite.entity.ts
