+++ src/modules/favorites/favorites.module.ts
@@ -5,7 +5,7 @@ import { FavoritesRepository } from './favorites.repository';
 import { TrackModule } from '../track/track.module';
 import { ArtistModule } from '../artist/artist.module';
 import { AlbumModule } from '../album/album.module';
-import { DatabaseModule } from '../database/database.module';
+import { DatabaseModule } from '../../database/database.module';
 
 @Module({
   imports: [
