+++ src/modules/favorites/favorites.service.ts
