+++ src/modules/track/track.module.ts
@@ -3,7 +3,7 @@ import { TrackController } from './track.controller';
 import { TrackService } from './track.service';
 import { TrackRepository } from './track.repository';
 import { FavoritesModule } from '../favorites/favorites.module';
-import { DatabaseModule } from '../database/database.module';
+import { DatabaseModule } from '../../database/database.module';
 
 @Module({
   imports: [forwardRef(() => FavoritesModule), DatabaseModule],
